refactor: Modularity for random.util
